+++ tests/src/test/java/org/eclipse/hono/tests/AmqpMessageContextConditionalVerifier.java
@@ -39,5 +39,4 @@ public final class AmqpMessageContextConditionalVerifier {
             assertThat(amqpMessageContext.getRawMessage().isDurable()).isTrue();
         }
     }
-
 }