+++ tests/src/test/java/org/eclipse/hono/tests/amqp/CommandAndControlAmqpIT.java
@@ -428,6 +428,7 @@ public class CommandAndControlAmqpIT extends AmqpAdapterTestBase {
                             ctx.verify(() -> {
                                 assertThat(response.getDeviceId()).isEqualTo(commandTargetDeviceId);
                                 assertThat(response.getTenantId()).isEqualTo(tenantId);
+                                assertThat(response.getCreationTime()).isNotNull();
                             });
                             return (Void) null;
                         });