+++ tests/src/test/java/org/eclipse/hono/tests/amqp/EventAmqpIT.java
@@ -66,6 +66,7 @@ public class EventAmqpIT extends AmqpUploadTestBase {
     @Override
     protected void assertAdditionalMessageProperties(final DownstreamMessage<? extends MessageContext> msg) {
         AmqpMessageContextConditionalVerifier.assertMessageIsDurable(msg);
+        assertThat(msg.getCreationTime()).isNotNull();
     }
 
     /**