+++ tests/src/test/java/org/eclipse/hono/tests/mqtt/CommandAndControlMqttIT.java
@@ -264,6 +264,7 @@ public class CommandAndControlMqttIT extends MqttTestBase {
                         ctx.verify(() -> {
                             assertThat(response.getDeviceId()).isEqualTo(commandTargetDeviceId);
                             assertThat(response.getTenantId()).isEqualTo(tenantId);
+                            assertThat(response.getCreationTime()).isNotNull();
                         });
                         return (Void) null;
                     });