+++ tests/src/test/java/org/eclipse/hono/tests/mqtt/EventMqttIT.java
@@ -101,6 +101,7 @@ public class EventMqttIT extends MqttPublishTestBase {
     @Override
     protected void assertAdditionalMessageProperties(final DownstreamMessage<? extends MessageContext> msg) {
         AmqpMessageContextConditionalVerifier.assertMessageIsDurable(msg);
+        assertThat(msg.getCreationTime()).isNotNull();
     }
 
     /**